requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
pandas==2.1.3
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-generativeai==0.3.1
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
python-multipart==0.0.6
pdfplumber==0.10.3
python-docx==1.1.0